# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Sample Functional DDD API
  description: 関数型プログラミングとDDDを組み合わせた受発注システムのAPI
  version: 1.0.0
  contact:
    name: API Support
  license:
    name: MIT

servers:
  - url: http://localhost:4000
    description: 開発環境

tags:
  - name: products
    description: 商品関連のAPI
  - name: orders
    description: 注文関連のAPI

paths:
  /api/products:
    get:
      summary: 商品一覧を取得
      description: 全ての商品の一覧を取得します
      operationId: getProducts
      tags:
        - products
      responses:
        '200':
          description: 商品一覧の取得に成功
          content:
            application/json:
              schema:
                type: object
                required:
                  - products
                properties:
                  products:
                    type: array
                    items:
                      $ref: '#/components/schemas/ProductDto'
              example:
                products:
                  - id: 1
                    title: iPhone 15
                    price: 999.99
                    description: 最新のiPhone
                  - id: 2
                    title: MacBook Pro
                    price: 2499.99
                    description: 高性能ノートPC
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/products/{id}:
    get:
      summary: 商品詳細を取得
      description: 指定されたIDの商品詳細を取得します
      operationId: getProductById
      tags:
        - products
      parameters:
        - name: id
          in: path
          required: true
          description: 商品ID
          schema:
            type: integer
            format: int32
            minimum: 1
          example: 1
      responses:
        '200':
          description: 商品詳細の取得に成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDto'
              example:
                id: 1
                title: iPhone 15
                price: 999.99
                description: 最新のiPhone
        '404':
          description: 商品が見つからない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  type: NOT_FOUND
                  message: 商品が見つかりません
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/orders:
    post:
      summary: 注文を作成
      description: 新しい注文を作成します
      operationId: createOrder
      tags:
        - orders
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
            example:
              orderItems:
                - productId: 1
                  quantity: 2
                - productId: 2
                  quantity: 1
              shippingAddress:
                postalCode: "100-0001"
                prefecture: "東京都"
                city: "千代田区"
                addressLine: "千代田1-1-1"
              customerInfo:
                name: "山田太郎"
                email: "taro@example.com"
                phone: "090-1234-5678"
      responses:
        '201':
          description: 注文の作成に成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDto'
        '400':
          description: バリデーションエラーまたは商品が見つからない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                validationError:
                  value:
                    error:
                      type: VALIDATION_ERROR
                      message: 郵便番号の形式が正しくありません
                productNotFound:
                  value:
                    error:
                      type: PRODUCT_NOT_FOUND
                      message: 商品ID 1 が見つかりません
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    ProductDto:
      type: object
      required:
        - id
        - title
        - price
      properties:
        id:
          type: integer
          format: int32
          minimum: 1
          description: 商品ID（正の整数）
          example: 1
        title:
          type: string
          minLength: 1
          description: 商品名
          example: iPhone 15
        price:
          type: number
          format: double
          minimum: 0
          description: 価格（0以上）
          example: 999.99
        description:
          type: string
          nullable: true
          description: 商品説明（任意）
          example: 最新のiPhone

    CreateOrderRequest:
      type: object
      required:
        - orderItems
        - shippingAddress
        - customerInfo
      properties:
        orderItems:
          type: array
          minItems: 1
          description: 注文明細のリスト（1件以上必須）
          items:
            type: object
            required:
              - productId
              - quantity
            properties:
              productId:
                type: integer
                format: int32
                minimum: 1
                description: 商品ID
                example: 1
              quantity:
                type: integer
                format: int32
                minimum: 1
                maximum: 100
                description: 数量（1〜100）
                example: 2
        shippingAddress:
          type: object
          required:
            - postalCode
            - prefecture
            - city
            - addressLine
          properties:
            postalCode:
              type: string
              pattern: '^\d{3}-\d{4}$'
              description: 郵便番号（XXX-XXXX形式）
              example: "100-0001"
            prefecture:
              type: string
              minLength: 1
              description: 都道府県
              example: "東京都"
            city:
              type: string
              minLength: 1
              description: 市区町村
              example: "千代田区"
            addressLine:
              type: string
              minLength: 1
              description: 番地・建物名
              example: "千代田1-1-1"
        customerInfo:
          type: object
          required:
            - name
            - email
            - phone
          properties:
            name:
              type: string
              minLength: 1
              description: 顧客名
              example: "山田太郎"
            email:
              type: string
              format: email
              description: メールアドレス
              example: "taro@example.com"
            phone:
              type: string
              pattern: '^\d{2,4}-\d{2,4}-\d{4}$'
              description: 電話番号（XX-XXXX-XXXX形式）
              example: "090-1234-5678"

    OrderDto:
      type: object
      required:
        - id
        - orderItems
        - shippingAddress
        - customerInfo
        - shippingStatus
        - totalAmount
        - createdAt
      properties:
        id:
          type: integer
          format: int32
          description: 注文ID
          example: 1
        orderItems:
          type: array
          items:
            type: object
            required:
              - productId
              - quantity
            properties:
              productId:
                type: integer
                format: int32
                description: 商品ID
                example: 1
              quantity:
                type: integer
                format: int32
                description: 数量
                example: 2
        shippingAddress:
          type: object
          required:
            - postalCode
            - prefecture
            - city
            - addressLine
          properties:
            postalCode:
              type: string
              description: 郵便番号
              example: "100-0001"
            prefecture:
              type: string
              description: 都道府県
              example: "東京都"
            city:
              type: string
              description: 市区町村
              example: "千代田区"
            addressLine:
              type: string
              description: 番地・建物名
              example: "千代田1-1-1"
        customerInfo:
          type: object
          required:
            - name
            - email
            - phone
          properties:
            name:
              type: string
              description: 顧客名
              example: "山田太郎"
            email:
              type: string
              description: メールアドレス
              example: "taro@example.com"
            phone:
              type: string
              description: 電話番号
              example: "090-1234-5678"
        shippingStatus:
          type: string
          enum: [PENDING, SHIPPED, DELIVERED]
          description: 配送ステータス
          example: PENDING
        totalAmount:
          type: number
          format: double
          minimum: 0
          description: 合計金額
          example: 2999.97
        createdAt:
          type: string
          format: date-time
          description: 作成日時
          example: "2024-01-01T00:00:00Z"

    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - type
            - message
          properties:
            type:
              type: string
              description: エラータイプ
              example: REPOSITORY_ERROR
            message:
              type: string
              description: エラーメッセージ
              example: データの取得に失敗しました